fix root file to clean
