+++ api/src/config/index.ts
@@ -1,0 +1,3 @@
+export * from './schema';
+export * from './database.config';
+export * from './associate';
