+++ api/src/routes/upload.ts
@@ -1,0 +1,6 @@
+import express from 'express';
+import { upload } from '../controller/upload';
+
+export const router = express.Router();
+
+router.get('/', upload);
